# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API untuk aplikasi Kasir menggunakan Clean Architecture.",
        "title": "Kasir API Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "crudapi-ku.up.railway.app",
    "basePath": "/",
    "paths": {
        "/api/produk": {
            "get": {
                "description": "Endpoint ini bisa menampilkan semua produk (GET) atau menambah produk baru (POST)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Produk"
                ],
                "summary": "List \u0026 Tambah Produk",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Product"
                            }
                        }
                    }
                }
            }
        },
        "/api/produk/{id}": {
            "put": {
                "description": "Mengubah data produk yang sudah ada berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Produk"
                ],
                "summary": "Update Produk",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data produk yang baru",
                        "name": "produk",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus data produk dari database berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Produk"
                ],
                "summary": "Hapus Produk",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Product deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "category_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                }
            }
        }
    }
}